mysiteapp/requirements.txt
appdirs==1.4.3
appnope==0.1.0
asgiref==3.2.5
astroid==2.3.3
backcall==0.1.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
decorator==4.4.2
distlib==0.3.0
dj-stripe==2.2.3
Django==3.0.4
django-filter==2.3.0
djangorestframework==3.11.1
filelock==3.0.12
idna==2.9
importlib-metadata==1.5.0
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.16.0
jsonfield==3.1.0
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
nose==1.3.7
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
pipenv==2020.8.13
prompt-toolkit==3.0.4
psycopg2==2.8.4
psycopg2-binary==2.8.4
ptyprocess==0.6.0
Pygments==2.6.1
pylint==2.4.4
python-termstyle==0.1.10
pytz==2019.3
requests==2.23.0
six==1.14.0
sniffer==0.4.1
sqlparse==0.3.1
stripe==2.43.0
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.8
virtualenv==20.0.10
virtualenv-clone==0.5.4
wcwidth==0.1.8
wrapt==1.11.2
zipp==3.1.0
